Use Debian 10.1 as base and target SDK 28

+++ library-rules/mpeg2dec.sh
@@ -1,8 +1,17 @@
 get_dependencies automake libtool
-do_fetch
+#do_fetch
 
-# mpeg2dec has an outdated autotools config which does not know of androideabi,
-# so replace it with a new one
+if [ ! -d mpeg2dec ]
+then
+    git clone https://salsa.debian.org/multimedia-team/mpeg2dec.git || exit 128
+fi
+
+cd mpeg2dec || exit 128
+
+patch -p1 < ../patch-mpeg2dec-confac.patch
+
+## mpeg2dec has an outdated autotools config which does not know of androideabi,
+## so replace it with a new one
 rm -rf .auto
 dh_update_autotools_config
 autoreconf -i
@@ -10,6 +19,8 @@ autoreconf -i
 # mpeg2dec includes non-PIC ARM assembly which is forbidden in Android 23+, so
 # disable it from being used
 sed -i 's/arm\*)/xxarmxx)/' configure
+
+chmod +x configure
 
 do_configure
 make -j$num_cpus -C libmpeg2 && \
